moves landing page content to .js
